freelance/pyproject.toml
[tool.poetry]
name = "freelance"
version = "0.1.0"
description = ""
authors = ["rezbuzzz <redbuzzzzzz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.6.0"
certifi = "2022.12.7"
cffi = "1.15.1"
charset-normalizer = "3.1.0"
cryptography = "40.0.2"
defusedxml = "0.7.1"
django = "4.2"
django-templated-mail = "1.1.1"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
djoser = "2.2.0"
idna = "3.4"
oauthlib = "3.2.2"
pillow = "9.5.0"
pycparser = "2.21"
pyjwt = "2.6.0"
python3-openid = "3.2.0"
pytz = "2023.3"
requests = "2.29.0"
requests-oauthlib = "1.3.1"
social-auth-app-django = "5.2.0"
social-auth-core = "4.4.2"
sqlparse = "0.4.4"
urllib3 = "1.26.15"
django-cors-headers = "3.7.0"
django-filter = "^23.2"
drf-yasg = "^1.21.5"
celery = "^5.2.7"
redis = "^4.5.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
